+++ track_database/tracks.py
@@ -17,6 +17,7 @@ available_tracks = [
     "fsds_default",
     "21_05_2023",
     "VSV",
+    "VSV_XS",
     "VSV_XL",
 ]